fix(testing): stabilize Event Hubs resume startup
